third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles-1/edit-inspector-stylesheet-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that adding a new rule creates inspector stylesheet resource and allows its live editing.

Inspector stylesheet URL: inspector-stylesheet
Inspector stylesheet content:
#inspected {}

Setting new content
Inspector stylesheet content:
#inspected { background-color: green; }
[expanded] 
element.style { ()

[expanded] 
#inspected { (inspector-stylesheet:1 -> inspector-stylesheet:1:13)
    background-color: green;

[expanded] 
div { (user agent stylesheet)
    display: block;
